# project1_backend/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "project1_backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Donhangs/ThanhToanQr": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "time",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/KiemtraThanhtoan": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "by",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "time",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { }
            }
          }
        }
      }
    },
    "/api/Donhangs/{id}": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Donhangs"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Donhang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/allOPofUser": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/allOFofUser": {
      "get": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/UpdateOrder": {
      "put": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderid",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "Status",
            "type": "string",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/HuyDon": {
      "post": {
        "tags": [
          "Donhangs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "orderid",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Donhangs/OrderProducts": {
      "post": {
        "tags": [
          "Donhangs"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "query",
            "name": "ship",
            "type": "integer",
            "format": "int32",
            "collectionFormat": "multi"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/InfoProduct"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Donhang"
            }
          }
        }
      }
    },
    "/api/Donhangs/OrderField": {
      "post": {
        "tags": [
          "Donhangs"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "phone",
            "type": "string",
            "collectionFormat": "multi"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SanbongDonhang"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Donhang"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Donhang": {
      "type": "object",
      "properties": {
        "orderid": {
          "format": "int32",
          "type": "integer"
        },
        "phonenumber": {
          "type": "string"
        },
        "totalcost": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "time": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "InfoProduct": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "price": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "color": {
          "type": "string"
        },
        "linkimg": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "SanbongDonhang": {
      "type": "object",
      "properties": {
        "orderid": {
          "format": "int32",
          "type": "integer"
        },
        "fieldid": {
          "type": "string"
        },
        "kip": {
          "format": "int32",
          "type": "integer"
        },
        "times": {
          "format": "date-time",
          "type": "string"
        },
        "cost": {
          "format": "int32",
          "type": "integer"
        },
        "note": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}